+++ src/main/java/view/OutputView.java
@@ -43,6 +43,6 @@ public class OutputView {
     }
 
     public static void printProfit(LottoProfit profit) {
-        System.out.println("총 수익률은 " + profit.getProfit() + "%입니다.");
+        System.out.println("총 수익률은 " + (long)profit.getProfit() + "%입니다.");
     }
 }
